apps/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
supabase>=2.3.0
python-dotenv>=1.0.0
